Avoid iterating over a dictionary
As a Dictionary in .NET is implemented as a hashtable, deleting a random element by
iterating N items to find the key to delete is a O(N) operation.  For the implementation
of a random cache eviction policy, just taking the first element of the dictionary is
random enough and saves a lot of time.

This change reduces the run time of a typical test run correlating 7 CSV files by 2.5x.