Use String.Split(char[]) instead of Regex.Split
The regular expression is only used to split on the comma and tab characters.
This is more efficiently done using the String.Split method.
The behaviour of only splitting into two fields, even if more than one separating
character is present is preserved.

The Using statement is just a nice way of ensuring a file gets closed in all cases.