Add caching of CSV files to binary serialized .bin files
As DateTime.TryParse is where most of the time is spent when using CSV files,
it makes sense to serialize the input data in a way that avoids parsing dates
over and over.

Serializing and deserialing the complete dictionary is too expensive,
because a lot of DictionaryEntry objects are created, which the
BinaryFormatter is not able to handle efficiently. So instead an array
of primitive type (Doubles) is stored on disk.